BitOperations: Aggressive Inlining should be for .NET 4.6.2 as well
Use `NETSTANDARD2_1 || NET462_OR_GREATER` for inlining. (Unnecessary) Guards were removed for `Span` and `ReadOnlySpan` API as these are only .NET Standard 2.1 anyway.

After reading the reference sources for `BitConverter`, we see that the `DangerousCopyXX` and `DangerousToXX` methods are checked for alignment. These could result in a crash on non-Intel machines. So the documentation was updated to reflect this. For safe methods, developers should use the `BitConverter` class instead.

Issue: DOTNET-871